Made implementation file for default factory functions Made implementation file for view factory
