+++ src/3space-studio/views/view_factory.hpp
@@ -15,43 +15,13 @@ using view_creator = graphics_view*(const shared::archive::file_info&, std::basi
 class view_factory
 {
 public:
-  void add_file_type(stream_validator* checker, view_creator* creator)
-  {
-    creators.emplace(checker, creator);
-  }
+  void add_file_type(stream_validator* checker, view_creator* creator);
 
-  void add_extension(std::string_view extension, stream_validator* checker)
-  {
-    validators.emplace(*extensions.emplace(extension).first, checker);
-  }
+  void add_extension(std::string_view extension, stream_validator* checker);
 
-  [[nodiscard]] std::vector<std::string_view> get_extensions() const
-  {
-    return std::vector<std::string_view>(extensions.cbegin(), extensions.cend());
-  }
+  [[nodiscard]] std::vector<std::string_view> get_extensions() const;
 
-  graphics_view* create_view(const shared::archive::file_info& file_info, std::basic_istream<std::byte>& stream, const studio::fs::file_system_archive& manager) const
-  {
-    auto archive_type = validators.equal_range(to_lower(file_info.filename.extension().string()));
-
-    for (auto it = archive_type.first; it != archive_type.second; ++it)
-    {
-      if (it->second(stream))
-      {
-        return creators.at(it->second)(file_info, stream, manager);
-      }
-    }
-
-    for (auto& [checker, creator] : creators)
-    {
-      if (checker(stream))
-      {
-        return creator(file_info, stream, manager);
-      }
-    }
-
-    return new default_view(file_info);
-  }
+  graphics_view* create_view(const shared::archive::file_info& file_info, std::basic_istream<std::byte>& stream, const studio::fs::file_system_archive& manager) const;
 
 private:
   std::set<std::string> extensions;
@@ -60,6 +30,6 @@ private:
   std::multimap<std::string_view, stream_validator*> validators;
 };
 
-view_factory create_default_view_factory();
+
 
 #endif//DARKSTARDTSCONVERTER_VIEW_FACTORY_HPP
